#655 and #658: consolidated Date-Calendar type checking code into one abstract class.
