+++ platform-extras/dolphin-platform-bean-validation/src/main/java/com/canoo/dp/impl/validation/PastPropertyValidator.java
@@ -1,7 +1,6 @@
 package com.canoo.dp.impl.validation;
 
 import javax.validation.ConstraintValidatorContext;
-import javax.validation.ValidationException;
 import javax.validation.constraints.Past;
 import java.util.Calendar;
 import java.util.Date;
@@ -9,33 +8,16 @@ import java.util.Date;
 /**
  * Validator that adds Dolphin Platform property support for the {@link Past} annotation.
  *
- * Note: using plain {@code Object} as data type because {@link Date} and {@link Calendar} have no common ancestor.
  */
-public final class PastPropertyValidator extends AbstractPropertyValidator<Past, Object> {
+public final class PastPropertyValidator extends AbstractDateAndCalendarValidator<Past> {
 
-    /**
-     * ctor
-     */
-    public PastPropertyValidator() {
-        super(Object.class);
+    @Override
+    protected boolean checkValidDate(Date date, ConstraintValidatorContext context) {
+        return date.before(new Date() );
     }
 
     @Override
-    protected boolean checkValid(Object value, ConstraintValidatorContext context) {
-        if (value instanceof Date) {
-            return checkValidDate((Date) value);
-        } else if (value instanceof Calendar) {
-            return checkValidCalendar((Calendar) value);
-        } else {
-            throw new ValidationException("Property contains value of type " + value.getClass() + " whereas only java.util.Date and java.util.Calendar are supported.");
-        }
-    }
-
-    private boolean checkValidCalendar(Calendar value) {
-        return value.before(Calendar.getInstance() );
-    }
-
-    private boolean checkValidDate(Date value) {
-        return value.before(new Date() );
+    protected boolean checkValidCalendar(Calendar calendar, ConstraintValidatorContext context) {
+        return calendar.before(Calendar.getInstance() );
     }
 }
